Add and Get Gender

+++ Rest-World-Disney-Service/infrastructure/driven-adapters/jpa-repository/src/main/java/co/com/disney/jpa/genderjpa/GenderRepositoryJPA.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GenderRepositoryJPA extends CrudRepository<GenderDataJPA, Integer> {
+public interface GenderRepositoryJPA extends CrudRepository<GenderDataJPA, Long> {
 	
 
 }
